backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.5.0
numpy>=1.21.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pydantic>=2.0.0
torchcam>=0.4.0
scikit-learn>=1.3.0
